+++ src/EventSeedWasFinished.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Turanct\Migraine;
+
+/**
+ * @psalm-immutable
+ */
+interface EventSeedWasFinished extends Event
+{
+    public function getConnectionString(): string;
+
+    public function getSeed(): string;
+}