+++ src/QueryFailed.php
@@ -15,6 +15,11 @@ final class QueryFailed extends \Exception
     /**
      * @var string
      */
+    private $seed;
+
+    /**
+     * @var string
+     */
     private $connection;
 
     public static function withMigrationData(
@@ -24,6 +29,18 @@ final class QueryFailed extends \Exception
     ): QueryFailed {
         $queryFailed = new static($message);
         $queryFailed->migration = $migration;
+        $queryFailed->connection = $connection;
+
+        return $queryFailed;
+    }
+
+    public static function withSeedData(
+        string $message,
+        string $seed = '',
+        string $connection = ''
+    ): QueryFailed {
+        $queryFailed = new static($message);
+        $queryFailed->seed = $seed;
         $queryFailed->connection = $connection;
 
         return $queryFailed;