+++ tests/fixtures/all-migrations-test/seeds/seed-2.sql
@@ -1,0 +1,3 @@
+UPDATE `test`
+SET name='This is a name'
+WHERE `id` = 'This is an id';