+++ tests/fixtures/single-migration-test/seeds/faulty-seed.sql
@@ -1,0 +1,1 @@
+INSRT INTO `test` (`id`) VALUES ('This is an id');